requirements.txt
aerich==0.8.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8
asyncpg==0.30.0
bcrypt==4.0.1
click==8.1.8
dictdiffer==0.9.0
environs==14.1.1
fastapi==0.115.12
h11==0.14.0
idna==3.10
iso8601==2.1.0
marshmallow==3.26.1
packaging==24.2
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.1
pydantic-core==2.33.0
pyjwt==2.10.1
pypika-tortoise==0.5.0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
sniffio==1.3.1
starlette==0.46.1
tomli-w==1.2.0
tomlkit==0.13.2
tortoise-orm==0.24.2
typing-extensions==4.13.0
typing-inspection==0.4.0
uvicorn==0.34.0
